reorganize contents related to bytes
