+++ grug/types/src/lib.rs
@@ -1,6 +1,7 @@
 mod address;
 mod app;
 mod bank;
+mod binary;
 mod bound;
 mod builder;
 mod bytes;
@@ -12,6 +13,8 @@ mod context;
 mod db;
 mod denom;
 mod empty;
+mod encoded_bytes;
+mod encoders;
 mod error;
 mod event;
 mod hash;
@@ -31,10 +34,10 @@ mod unique_vec;
 mod utils;
 
 pub use {
-    address::*, app::*, bank::*, bound::*, builder::*, bytes::*, changeset::*, coin::*,
-    coin_pair::*, coins::*, context::*, db::*, denom::*, empty::*, error::*, event::*, hash::*,
-    hashers::*, imports::*, non_empty::*, non_zero::*, query::*, response::*, result::*,
-    serializers::*, signer::*, time::*, tx::*, unique_vec::*, utils::*,
+    address::*, app::*, bank::*, binary::*, bound::*, builder::*, bytes::*, changeset::*, coin::*,
+    coin_pair::*, coins::*, context::*, db::*, denom::*, empty::*, encoded_bytes::*, encoders::*,
+    error::*, event::*, hash::*, hashers::*, imports::*, non_empty::*, non_zero::*, query::*,
+    response::*, result::*, serializers::*, signer::*, time::*, tx::*, unique_vec::*, utils::*,
 };
 
 // ---------------------------------- testing ----------------------------------
